fresh start
i'm going to try rewriting charm to be a library purely for rendering - no mouse events or state.

i've been trying to make charm a low-level GUI lib as well as a layout lib because i feel like it's a missed opportunity if i don't. but this isn't actually a use case that i'll use, and i think it shows in the design. the library is overcomplicated and doesn't solve the problems it sets out to.

i hope that by rewriting charm with a more singular focus, it'll benefit in design and functionality.