add :shift and :wrap
i think the wrap function has a lot of potential for improvement. it's not very flexible, and i think a lot of what it does could be separated into standalone functions. but i'm trying to keep the API minimal, so i'm going to make improvements as i run into annoying situations, not preemptively