fix transform bug introduced by removing render
explanation:

when children are added to a transform element or the transformation is changed, the transform element recalculates its dimensions. it will move to surround the newly transformed children. we don't want this movement to affect how the children are drawn, so we set the childrenShift variables, which are used later to translate the children in the draw function.

i was wondering why i had to compensate for the transform element's movement twice - i found out why!

the transform's dimensions were adjusted whenever onEndChildren was called. the render phase of drawing was calling beginChildren and endChildren on each element, triggering that callback. so transform elements were adjusting their position twice instead of just once. now that i've removed the render phase, this calculation can go back to the way it should be~