don't allow transforms to move their children
it seems difficult, if not impossible, to make a Transform element that both properly reflects the bounds of its children and lets you change the position of the children. from my experience so far, these are conflicting behaviors that cause a lot of problems.

so i'm going to go back to my original design intent, which is that the transform element *only* reflects the bounds of the children, but cannot visually change them. this does lead to some limitations to how you can use transforms, but i suspect that none of them will be too hard to deal with. we'll find out!