maintain the visual position of children regardless of origin
i figure that since you can't move a transform directly, it might be tricky to position transformed elements properly, and it would be less tricky if changing the origin of a transformation didn't also move the children. so i'm automatically moving the transformation to maintain the same visual position on screen after setting the origin. hopefully this is a sensible default.